Intial Commit
The test.py file is for experimentaiton with the poorly documented
AppIndicator3 interface. I've managed to capure most of the main
functionality with this experimentation by this point.

The unified-adgenda.py file contains the actual content of the
project. Currently, it is capable of reading arbitrary JSON
formatted "settings" from a file in ~/.config, and fetching files
from urls specified in the config file.

The README.md file primarily contians the TODO list, where
completed items have strikethrough decoration.